Separated change password modal and used custom hook for processing modal
