+++ src/containers/edit-profile/password-security-tab/PasswordSecurityTab.styles.ts
@@ -38,7 +38,7 @@ export const styles = {
     display: 'flex',
     alignItems: 'center',
     justifyContent: 'center',
-    height: '100vh',
+    width: '700px',
     position: 'relative'
   }
 }
